tests: Move 2 aggregator RPC tests to aggregator unit tests.
